+++ constructorabackend/src/main/java/constructora/constructorabackend/service/ProveedorService.java
@@ -15,10 +15,19 @@ public class ProveedorService implements IProveedorService{
     private IProveedorRepository iProveedorRepository;
 
     @Override
-    public ProveedorModel saveProveedor(ProveedorModel proveedorModel) { return iProveedorRepository.save(proveedorModel); }
+    public ProveedorModel saveProveedor(ProveedorModel proveedorModel) {
+        validateProveedor(proveedorModel);
+        return iProveedorRepository.save(proveedorModel);
+    }
 
     @Override
     public ProveedorModel updateProveedor(ProveedorModel proveedorModel) {
+
+        if (!iProveedorRepository.findById(proveedorModel.getIdProveedores()).isPresent()) {
+            throw new IllegalArgumentException("Proveedor no encontrado");
+        }
+        validateProveedor(proveedorModel);
+
         return iProveedorRepository.save(proveedorModel);
     }
 
@@ -29,11 +38,27 @@ public class ProveedorService implements IProveedorService{
 
     @Override
     public Optional<ProveedorModel> getProveedorById(Integer id) {
+        if (id <= 0) {
+            throw new IllegalArgumentException("ID no válido");
+        }
         return iProveedorRepository.findById(id);
     }
 
     @Override
     public void deleteProveedor(Integer id) {
+        if (id <= 0) {
+            throw new IllegalArgumentException("ID no válido");
+        }
         iProveedorRepository.deleteById(id);
+    }
+
+
+    private void validateProveedor(ProveedorModel proveedor) {
+        if (proveedor.getNombreProveedor() == null || proveedor.getNombreProveedor().isEmpty() ||
+                proveedor.getTipoServicio() == null || proveedor.getTipoServicio().isEmpty() ||
+                proveedor.getCorreo() == null || !proveedor.getCorreo().contains("@") ||
+                proveedor.getTelefono() <= 0) {
+            throw new IllegalArgumentException("Datos inválidos");
+        }
     }
 }
